Keep simulators happy.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@21403 a1c6a512-1295-4272-9138-f99709370657

+++ apps/tagcache.h
@@ -173,7 +173,7 @@ struct tagcache_search {
     int clause_count;
     int list_position;
     int seek_pos;
-    int32_t position;
+    long position;
     int entry_count;
     bool valid;
     bool initialized;
